+++ CHANGELOG.md
@@ -14,6 +14,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Display error message if attachment upload fails with code `413 Request Entity Too Large`.
 
+### Removed
+
+- Removed EIP-191 activity signing.
+
 ### Fixed
 
 - Don't save drafts when editing post.